refine system file structure
