+++ src/utils/registry.rs
